+++ tests/browser/find_by_text.test.js
@@ -85,11 +85,16 @@ describe("Find By Text", function() {
         }, "Error: [findByTextContaining] Invalid selector.");
     });
 
-    // Tests #342
-    it.skip("Find By Text With Quotes Character", async() => {
+    it("Find By Text With Quotes Character", async() => {
         await browser.open(configUrls.weirdText);
         const elems = await browser.findByText("\"quotes'");
         assert.strictEqual(elems.length, 1);
+    });
+
+    it("Find By Text With Multiple Quotes Character Fail", async() => {
+        await browser.open(configUrls.weirdText);
+        const elems = await browser.findByText("\"quo'tes''");
+        assert.strictEqual(elems.length, 0);
     });
 
     it("Find By Text With Quotes Character Fails", async() => {
@@ -104,8 +109,7 @@ describe("Find By Text", function() {
         assert.strictEqual(elems.length, 1);
     });
 
-    // Tests #342
-    it.skip("Find By Text Containing With Simple Quote Character", async() => {
+    it("Find By Text Containing With Simple Quote Character", async() => {
         await browser.open(configUrls.weirdText);
         const elems = await browser.findByTextContaining("es'");
         assert.strictEqual(elems.length, 1);